Update to reference VS 2019 build tools. (#252)
Updates the SDK Dockerfiles to reference the VS 2019 build tools. This basically consists of updates to paths and some ngen workarounds.

Also includes some changes to delete the VS Packages folder for SDK images to save on space of the image.

windowservercore-1709 images have not been updated since that is going to be EOL and want to avoid any potential breakage there.