Fix PR build runs (#280)
The PR build is having issues completing in a reasonable amount of time. I've trimmed the test matrix a bit to cut down on the number of images that need to be built in a given leg but still provide decent enough coverage of runtime/sdk combos on the whole.

I also found that there was an issue in the definition of the versionFilter parameter for tests that was causing the 4.7 leg to attempt to run tests for 4.7.1 and 4.7.2. This is due to the '*' character that was appended to the 4.7 string causing it to match those other versions. I've removed it so that it will only match on the exact version string.