Fix broken "Dockerfiles" search link
The recent .NET 10 PR (#1271) has a GitHub search link that is reporting a 429 in reviewdog; on closer inspection, I found that it no longer return results, but rather a warning.

I have replaced it with the equivalent search link, but with guidance from #1175, which fixed a similar link but restricted it to `src/` Dockerfiles.

Note this may not necessarily fix the 429 that reviewdog is getting as there may be other factors causing it.